+++ resources/views/hour/index.blade.php
@@ -19,6 +19,7 @@
             <th></th>
             <th sortId="time">{{__('Time')}}</th>
             <th sortId="duration" class="lkn-align-right">{{__('Duration')}}</th>
+            <th sortId="factor" class="lkn-align-right">{{__('Factor')}}</th>
             <th sortId="hourtype">{{__('Hourtype')}}</th>
             <th sortId="hourstate">{{__('Hourstate')}}</th>
             <th sortId="description">{{__('Description')}}</th>
@@ -32,6 +33,7 @@
             <td><x-laraknife.icons.change-record module="hour" no="{{ $hour->id }}" /></td>
               <td>{{$context->asDateString($hour->time)}}</td>
               <td class="lkn-align-right">{{$context->asDuration($hour->duration)}}</td>
+              <td class="lkn-align-right">{{$hour->factor}}</td>
               <td> {{ __($hour->hourtype) }}</td>
               <td> {{ __($hour->hourstate) }}</td>
               <td>{{$hour->description}}</td>